backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
openai==1.58.1
anthropic==0.42.0
google-genai==1.12.1
reportlab==4.2.5
aiosqlite==0.20.0
greenlet==3.3.2
pydantic==2.10.4
